+++ bip_utils/bip/bip39/bip39_mnemonic.py
@@ -27,8 +27,8 @@ from enum import auto, Enum, IntEnum, unique
 from typing import Dict, List, Optional, Union
 from bip_utils.bip.bip39.bip39_ex import Bip39ChecksumError
 from bip_utils.bip.bip39.bip39_entropy_generator import Bip39EntropyGenerator
-from bip_utils.utils import (
-    ConvUtils, CryptoUtils,
+from bip_utils.utils import ConvUtils, CryptoUtils
+from bip_utils.utils.mnemonic import (
     Mnemonic, MnemonicWordsList, MnemonicWordsListGetterBase
 )
