+++ bip_utils/bip/conf/bip84/bip84_conf.py
@@ -22,7 +22,7 @@
 # Imports
 from bip_utils.bip.bip32 import Bip32KeyNetVersions, Bip32Secp256k1
 from bip_utils.bip.conf.common import *
-from bip_utils.utils import CoinNames
+from bip_utils.utils.conf import CoinNames
 
 
 # Bitcoin key net version (zpub / zprv)
