+++ bip_utils/monero/mnemonic/monero_entropy_generator.py
@@ -22,7 +22,7 @@
 # Imports
 from enum import IntEnum, unique
 from typing import List, Union
-from bip_utils.utils import EntropyGenerator
+from bip_utils.utils.mnemonic import EntropyGenerator
 
 
 @unique
