+++ bip_utils/substrate/__init__.py
@@ -1,5 +1,4 @@
 from bip_utils.substrate.substrate_ex import SubstrateKeyError, SubstratePathError
-from bip_utils.substrate.substrate_bip39 import SubstrateBip39SeedGenerator
 from bip_utils.substrate.substrate_keys import SubstratePublicKey, SubstratePrivateKey
 from bip_utils.substrate.substrate_path import SubstratePathElem, SubstratePath, SubstratePathParser
 from bip_utils.substrate.substrate import SubstrateCoins, Substrate
