+++ bip_utils/substrate/conf/substrate_coin_conf.py
@@ -20,7 +20,7 @@
 
 
 # Imports
-from bip_utils.utils import CoinNames
+from bip_utils.utils.conf import CoinNames
 
 
 class SubstrateCoinConf:
