+++ bip_utils/utils/__init__.py
@@ -1,7 +1,3 @@
-from bip_utils.utils.conf import CoinNames
 from bip_utils.utils.misc import (
     AlgoUtils, Base32Encoder, BitUtils, ConvUtils, CryptoUtils, DataBytes
-)
-from bip_utils.utils.mnemonic import (
-    EntropyGenerator, Mnemonic, MnemonicWordsList, MnemonicWordsListFileReader, MnemonicWordsListGetterBase
 )
